deparate the logs from the mane table and save current changes
